[TASK] Moving backend related classes into proper `Backend` namespace
