+++ Classes/Backend/Configuration/Wizard/NewContentElementWizard.php
